# api-gestao-pro/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API Gest?o Pro
  version: 1.0.0
servers:
  - url: http://localhost:4002
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/auth/register:
    post:
      summary: Registar utilizador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        '201': { description: Created }
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Token }
  /api/projects:
    get:
      summary: Listar projetos
      responses:
        '200': { description: OK }
    post:
      summary: Criar projeto
      responses:
        '201': { description: Created }
  /api/projects/{id}:
    put:
      summary: Atualizar projeto
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Remover projeto
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
  /api/projects/{id}/tasks:
    get:
      summary: Listar tarefas do projeto
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Criar tarefa
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '201': { description: Created }
  /api/tasks/{id}:
    put:
      summary: Atualizar tarefa
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Remover tarefa
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
  /api/tasks/{id}/status:
    patch:
      summary: Atualizar status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }